Refactor: Modal 관련 컴포넌트 정리
Modal, LoginModal 정리
